backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["zart227 <zart227@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
django-rest-auth = "^0.9.5"
djoser = "^2.3.1"
drf-yasg = "^1.21.8"
django-celery-beat = "^2.7.0"
django-jazzmin = "^3.0.1"
django-admin-soft-dashboard = "^1.0.24"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
dj-rest-auth = "^7.0.0"
django-cors-headers = "^4.6.0"
redis = "^5.2.1"
django-celery-results = "^2.5.1"
rest-framework-simplejwt = "^0.0.2"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
black = "^24.10.0"
flake8 = "^7.1.1"
djlint = "^1.36.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
